Backend/requirements.txt
asgiref==3.9.2
Django==5.2.6
sqlparse==0.5.3
chromadb==1.1.0
firecrawl-py==4.3.6
openai==1.109.1
weasyprint==66.0
python-decouple==3.8
google-generativeai==0.8.5